# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "MojoPedia API",
        "description": "REST API untuk sistem manajemen pengetahuan MojoPedia. Menyediakan endpoint untuk manajemen user, role, dan integrasi Keycloak SSO.",
        "contact": {
            "name": "Tim Pengembang MojoPedia",
            "email": "admin@mojopedia.go.id"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/categories/{id}": {
            "get": {
                "tags": [
                    "Categories"
                ],
                "summary": "Detail satu kategori",
                "operationId": "getCategoryDetail",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detail kategori ditemukan"
                    },
                    "404": {
                        "description": "Kategori tidak ditemukan"
                    }
                }
            }
        },
        "/admin/categories": {
            "post": {
                "tags": [
                    "Categories"
                ],
                "summary": "Tambah kategori baru (Admin)",
                "operationId": "createCategory",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nama_kategori"
                                ],
                                "properties": {
                                    "nama_kategori": {
                                        "type": "string",
                                        "example": "Informatika & Siber"
                                    },
                                    "deskripsi": {
                                        "type": "string",
                                        "example": "Dokumen seputar infrastruktur IT & keamanan informasi"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Kategori berhasil ditambahkan"
                    },
                    "422": {
                        "description": "Validasi gagal"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/admin/categories/{id}": {
            "put": {
                "tags": [
                    "Categories"
                ],
                "summary": "Perbarui kategori (Admin)",
                "operationId": "updateCategory",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nama_kategori"
                                ],
                                "properties": {
                                    "nama_kategori": {
                                        "type": "string",
                                        "example": "Informatika"
                                    },
                                    "deskripsi": {
                                        "type": "string",
                                        "example": "Deskripsi baru"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Kategori berhasil diperbarui"
                    },
                    "404": {
                        "description": "Kategori tidak ditemukan"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Categories"
                ],
                "summary": "Hapus kategori (Admin)",
                "operationId": "deleteCategory",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Kategori berhasil dihapus"
                    },
                    "400": {
                        "description": "Kategori masih memiliki pengetahuan terkait"
                    },
                    "404": {
                        "description": "Kategori tidak ditemukan"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Masukkan token Sanctum Anda. Dapatkan token melalui endpoint login.",
                "bearerFormat": "Token",
                "scheme": "bearer"
            }
        }
    },
...
}